management-accounting/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg[binary]==3.1.18
alembic==1.13.1
pydantic-settings==2.3.4
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
itsdangerous==2.2.0
orjson==3.10.6
redis==5.0.4
celery==5.3.6
pdfplumber==0.11.0
tabula-py==2.9.0
pandas==2.2.2
openpyxl==3.1.2
python-json-logger==2.0.7
pytest==8.2.0
reportlab==4.2.0
